last version

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: oumimoun <oumimoun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/01 18:40:31 by oumimoun          #+#    #+#             */
-/*   Updated: 2023/11/20 14:48:59 by oumimoun         ###   ########.fr       */
+/*   Updated: 2023/11/30 11:13:39 by oumimoun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <stddef.h>
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	int	i;
